qemu, ciao-launcher:  Move ConfigDrive ISO creation code to qemu
Launcher's ConfigDrive ISO creation function, createCloudInitISO has
been moved to the qemu package so that it can be re-used by ciao-down.
